server/requirements.txt
asgiref==3.8.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Django==5.1.2
django-cors-headers==4.4.0
django-ratelimit==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Flask==3.0.3
Flask-Cors==5.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
pillow==10.4.0
pipedrive-python-lib==1.2.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.9.0
pyotp==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.4
